Implement container-progress() function
As introduced in https://drafts.csswg.org/css-values-5/#container-progress-func
The container-progress() functional notation returns a <number> value
representing current value of the size-feature of some container, that
can be either named or unnamed, as a progress value between two explicit
calc values.

Note: for now we only support width and height features.

Note: now it's an editor's draft with a very strong chances to be accepted.
Bug: 40944203
Change-Id: I4664ad8d01174bfdc9621ca304caecc9db37c9ca