Implement -internal-auto-bikeshed()
This patch implements -internal-auto-bikeshed(), a CSS function that can
be used in UA stylesheets which takes two values. If the computed style
of the element is not appearance:bikeshed, then the first value is used.
Otherwise, the second value is used.

This patch only uses it for background-color on <select>, but I will try
using it for additional properties like border next.

Bug: 1511354
Change-Id: I70fea56ef8a8c931418872039f1df38e5324f55f